improving the renderer, refactoring multiple ways to display grids
